Improve MapboxNavigation project structure.
