+++ MapboxNavigation/AVAudioSession.swift
@@ -1,4 +1,3 @@
-
 import AVFoundation
 
 public extension AVAudioSession {
