+++ MapboxNavigation/CarPlaySearchController+CPSearchTemplateDelegate.swift
