+++ MapboxNavigation/MapboxSpeechSynthesizer.swift
@@ -1,4 +1,3 @@
-
 import AVFoundation
 import MapboxCoreNavigation
 import MapboxDirections
