+++ MapboxNavigation/NavigationView.swift
@@ -1,5 +1,4 @@
 import UIKit
-import MapboxDirections
 
 /**
  A view that represents the root view of the MapboxNavigation drop-in UI.
