+++ MapboxNavigation/SystemSpeechSynthesizer.swift
@@ -1,4 +1,3 @@
-
 import AVFoundation
 import MapboxDirections
 import MapboxCoreNavigation
