+++ MapboxNavigation/UIEdgeInsets.swift
@@ -1,6 +1,5 @@
 import Foundation
 import UIKit
-import MapboxDirections
 
 extension UIEdgeInsets {
     public static func +(left: UIEdgeInsets, right: UIEdgeInsets) -> UIEdgeInsets {
