+++ MapboxNavigation/UserCourseView.swift
@@ -1,6 +1,4 @@
 import UIKit
-import Turf
-import Mapbox
 
 /**
  A protocol that represents a `UIView` which tracks the user’s location and course on a `NavigationMapView`.
